requirements.txt
beautifulsoup4>=4.12
requests>=2.31
python-dateutil>=2.8
python-dotenv>=1.0
duckdb>=0.10
streamlit>=1.33
firecrawl-py>=0.0.20
anthropic>=0.30
